requirements.txt
cython
setuptools
datasets<4.0.0
simplemma
lxml
hf_xet
tqdm
nltk
python-dotenv
numpy
